+++ .coveragerc
@@ -2,12 +2,12 @@
 parallel = True
 branch = True
 source =
-    aqx
+    braket
 
 [paths]
 source =
-    src/aqx
-    .tox/*/lib/python*/site-packages/aqx
+    src/braket
+    .tox/*/lib/python*/site-packages/braket
 
 [report]
 show_missing = True
