+++ doc/index.rst
@@ -1,9 +1,9 @@
-Amazon Qx Python IR
+Amazon Braket Python IR
 ====================
 
-Amazon Qx Python IR is an open source library that contains all the intermediate representations (IR) for AQx quantum tasks and offers serialization and deserialization of those IR payloads. Think of the IR as the contract between the AQx SDK and AQx API for quantum programs.
+Amazon Braket Python IR is an open source library that contains all the intermediate representations (IR) for Amazon Braket quantum tasks and offers serialization and deserialization of those IR payloads. Think of the IR as the contract between the Amazon Braket SDK and Amazon Braket API for quantum programs.
 
-Here you'll find an overview and API documentation for Amazon Qx Python IR. The project homepage is in GitHub, https://github.com/aws/aqx-python-ir, where you can find the source and installation instructions for the library.
+Here you'll find an overview and API documentation for Amazon Braket Python IR. The project homepage is in GitHub, https://github.com/aws/braket-python-ir, where you can find the source and installation instructions for the library.
 
 Indices and tables
 __________________
