+++ src/braket/ir/jaqcd/__init__.py
@@ -1,0 +1,46 @@
+# Copyright 2019-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You
+# may not use this file except in compliance with the License. A copy of
+# the License is located at
+#
+#     http://aws.amazon.com/apache2.0/
+#
+# or in the "license" file accompanying this file. This file is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
+# ANY KIND, either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+
+from braket.ir.jaqcd.instructions import (  # noqa: F401
+    CY,
+    CZ,
+    XX,
+    YY,
+    ZZ,
+    CCNot,
+    CNot,
+    CPhaseShift,
+    CPhaseShift00,
+    CPhaseShift01,
+    CPhaseShift10,
+    CSwap,
+    H,
+    I,
+    ISwap,
+    PhaseShift,
+    PSwap,
+    Rx,
+    Ry,
+    Rz,
+    S,
+    Si,
+    Swap,
+    T,
+    Ti,
+    V,
+    Vi,
+    X,
+    Y,
+    Z,
+)
+from braket.ir.jaqcd.program import Program  # noqa: F401
