+++ tox.ini
@@ -55,7 +55,7 @@ deps =
     sphinx
     sphinx-rtd-theme
 commands =
-    sphinx-apidoc -e -f -M --implicit-namespaces -H "API Reference" -o doc/_apidoc src/aqx
+    sphinx-apidoc -e -f -M --implicit-namespaces -H "API Reference" -o doc/_apidoc src/braket
     sphinx-build -E -T -b html doc build/documentation/html
 
 [testenv:serve-docs]
